backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
httpx==0.27.2
apscheduler==3.10.4
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
